SEQNG-703 Use Monocle lenses to create diode ModelRW
