+++ modules/seqexec/web/client/src/main/scala/seqexec/web/client/model/model.scala
@@ -8,15 +8,11 @@ import cats.implicits._
 import diode.data.Pot
 import gem.Observation
 import gem.enum.Site
-import monocle.Getter
 import monocle.macros.Lenses
 import org.scalajs.dom.WebSocket
-import seqexec.model.{ ClientID, Conditions, Observer, UserDetails, SequencesQueue, SequenceView }
+import seqexec.model.ClientID
 import seqexec.model.events._
 import seqexec.web.common.FixedLengthBuffer
-import seqexec.web.client.components.sequence.steps.StepConfigTable
-import seqexec.web.client.components.QueueTableBody
-import web.client.table._
 
 final case class RunningStep(last: Int, total: Int)
 
@@ -60,48 +56,12 @@ final case class GlobalLog(log: FixedLengthBuffer[ServerLogMessage], display: Se
 final case class ResourcesConflict(visibility: SectionVisibilityState, id: Option[Observation.Id])
 
 /**
- * UI model, changes here will update the UI
- */
-@Lenses
-final case class SeqexecUIModel(navLocation: Pages.SeqexecPages,
-                          user: Option[UserDetails],
-                          sequences: SequencesQueue[SequenceView],
-                          loginBox: SectionVisibilityState,
-                          resourceConflict: ResourcesConflict,
-                          globalLog: GlobalLog,
-                          sequencesOnDisplay: SequencesOnDisplay,
-                          syncInProgress: Boolean,
-                          configTableState: TableState[StepConfigTable.TableColumn],
-                          queueTableState: TableState[QueueTableBody.TableColumn],
-                          defaultObserver: Observer,
-                          firstLoad: Boolean)
-
-@SuppressWarnings(Array("org.wartremover.warts.PublicInference"))
-object SeqexecUIModel {
-  val noSequencesLoaded: SequencesQueue[SequenceView] = SequencesQueue[SequenceView](Map.empty, Conditions.Default, None, Nil)
-  val Initial: SeqexecUIModel = SeqexecUIModel(
-    Pages.Root,
-    None,
-    noSequencesLoaded,
-    SectionClosed,
-    ResourcesConflict(SectionClosed, None),
-    GlobalLog(FixedLengthBuffer.unsafeFromInt(500), SectionClosed),
-    SequencesOnDisplay.empty,
-    syncInProgress = false,
-    StepConfigTable.InitialTableState,
-    QueueTableBody.InitialTableState.tableState,
-    Observer(""),
-    firstLoad = true)
-
-  def sequenceReader(id: Observation.Id): Getter[SeqexecUIModel, Option[SequenceView]] =
-    SeqexecUIModel.sequences composeGetter SequencesQueue.queueItemG[SequenceView](_.id === id)
-  }
-
-/**
   * Root of the UI Model of the application
   */
+@Lenses
 final case class SeqexecAppRootModel(ws: WebSocketConnection, site: Option[Site], clientId: Option[ClientID], uiModel: SeqexecUIModel)
 
+@SuppressWarnings(Array("org.wartremover.warts.PublicInference"))
 object SeqexecAppRootModel {
   val Initial: SeqexecAppRootModel = SeqexecAppRootModel(WebSocketConnection.Empty, None, None, SeqexecUIModel.Initial)
 }
